requirements.txt
django==5.1.4
psycopg2==2.9.10
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt
drf-nested-routers==0.94.1
pytest-django==4.9.0
django-cors-headers==4.6.0
pillow==11.1.0